[elixir-2025] upgrade elixir to 1.17.3
the relevant change here is bumping cowboy to 2.9, the tests pass with
elixir 1.17.3 and erlang 25, going to erlang 26 causes connection to
timeout so I am sticking to erlang 25 as in the `Dockerfile.elixir` file
for now.

Also had to update the SSL certs since the CA had expired Nov 8th 2024.